+++ radix-engine-queries/src/typed_substate_layout.rs
@@ -13,10 +13,10 @@ pub use radix_engine::blueprints::pool::one_resource_pool;
 pub use radix_engine::blueprints::pool::two_resource_pool;
 pub use radix_engine::blueprints::resource::*;
 pub use radix_engine::blueprints::transaction_tracker::*;
-pub use radix_engine::system::node_module::metadata::*;
-pub use radix_engine::system::node_module::role_assignment::*;
-pub use radix_engine::system::node_module::royalty::*;
-pub use radix_engine::system::node_module::type_info::*;
+pub use radix_engine::system::node_modules::metadata::*;
+pub use radix_engine::system::node_modules::role_assignment::*;
+pub use radix_engine::system::node_modules::royalty::*;
+pub use radix_engine::system::node_modules::type_info::*;
 use radix_engine::system::system::FieldSubstate;
 pub use radix_engine::system::system::KeyValueEntrySubstate;
 pub use radix_engine_interface::api::node_modules::royalty::*;
