+++ radix-engine-tests/tests/royalty.rs
@@ -1,6 +1,6 @@
 use radix_engine::blueprints::package::PackageError;
 use radix_engine::errors::{ApplicationError, RuntimeError, SystemError};
-use radix_engine::system::node_module::royalty::ComponentRoyaltyError;
+use radix_engine::system::node_modules::royalty::ComponentRoyaltyError;
 use radix_engine::types::*;
 use radix_engine_interface::blueprints::resource::FromPublicKey;
 use scrypto_unit::*;
