+++ radix-engine-tests/tests/system_access_rule.rs
@@ -1,7 +1,7 @@
 use native_sdk::modules::role_assignment::{RoleAssignment, RoleAssignmentObject};
 use radix_engine::errors::{ApplicationError, RuntimeError};
 use radix_engine::kernel::kernel_api::{KernelNodeApi, KernelSubstateApi};
-use radix_engine::system::node_module::role_assignment::RoleAssignmentError;
+use radix_engine::system::node_modules::role_assignment::RoleAssignmentError;
 use radix_engine::system::system_callback::SystemLockData;
 use radix_engine::types::*;
 use radix_engine::vm::{OverridePackageCode, VmInvoke};
