+++ radix-engine/src/system/mod.rs
@@ -2,7 +2,7 @@ pub mod bootstrap;
 pub mod id_allocation;
 pub mod module;
 pub mod node_init;
-pub mod node_module;
+pub mod node_modules;
 pub mod payload_validation;
 pub mod system;
 pub mod system_callback;
