+++ radix-engine/src/system/payload_validation.rs
@@ -9,7 +9,7 @@ use radix_engine_interface::constants::*;
 use sbor::rust::prelude::*;
 use sbor::traversal::TerminalValueRef;
 
-use super::node_module::type_info::TypeInfoSubstate;
+use super::node_modules::type_info::TypeInfoSubstate;
 use super::system::SystemService;
 use super::system_callback::SystemConfig;
 use super::system_callback_api::SystemCallbackObject;
