+++ radix-engine/src/system/system_callback.rs
@@ -1,4 +1,4 @@
-use super::node_module::type_info::{TypeInfoBlueprint, TypeInfoSubstate};
+use super::node_modules::type_info::{TypeInfoBlueprint, TypeInfoSubstate};
 use crate::blueprints::account::ACCOUNT_CREATE_VIRTUAL_ED25519_ID;
 use crate::blueprints::account::ACCOUNT_CREATE_VIRTUAL_SECP256K1_ID;
 use crate::blueprints::identity::IDENTITY_CREATE_VIRTUAL_ED25519_ID;
