+++ radix-engine/src/system/system_db_reader.rs
@@ -12,7 +12,7 @@ use radix_engine_store_interface::{
 };
 use sbor::rust::prelude::*;
 
-use crate::system::node_module::type_info::TypeInfoSubstate;
+use crate::system::node_modules::type_info::TypeInfoSubstate;
 use crate::system::system::KeyValueEntrySubstate;
 use crate::track::TrackedNode;
