+++ radix-engine/src/system/system_modules/auth/README.md
@@ -19,7 +19,7 @@ Access Control** (RBAC) and **Attribute-Based Access Control** (ABAC) techniques
 On the callee side, instead of roles being assigned to a "user" (of which there is no concept in our decentralized
 ledger), roles are assigned through resource ownership. For example, a "Staff" role could be assigned to anyone who
 can show proof that they own a "Staff" resource token. This is defined at the callee's Object instantiation
-through the [RoleAssignment](../../node_module/role_assignment) object module.
+through the [RoleAssignment](../../node_modules/role_assignment) object module.
 
 ## Role Definition
