+++ radix-engine/src/system/system_modules/auth/auth_module.rs
@@ -7,8 +7,8 @@ use crate::kernel::call_frame::ReferenceOrigin;
 use crate::kernel::kernel_api::{KernelApi, KernelInternalApi, KernelNodeApi, KernelSubstateApi};
 use crate::system::module::KernelModule;
 use crate::system::node_init::type_info_partition;
-use crate::system::node_module::role_assignment::RoleAssignmentNativePackage;
-use crate::system::node_module::type_info::TypeInfoSubstate;
+use crate::system::node_modules::role_assignment::RoleAssignmentNativePackage;
+use crate::system::node_modules::type_info::TypeInfoSubstate;
 use crate::system::system::{FieldSubstate, SystemService};
 use crate::system::system_callback::{SystemConfig, SystemLockData};
 use crate::system::system_callback_api::SystemCallbackObject;
