+++ radix-engine/src/transaction/state_update_summary.rs
@@ -9,7 +9,7 @@ use radix_engine_store_interface::{
 };
 use sbor::rust::prelude::*;
 
-use crate::system::node_module::type_info::TypeInfoSubstate;
+use crate::system::node_modules::type_info::TypeInfoSubstate;
 use crate::system::system::FieldSubstate;
 use crate::system::system_db_reader::SystemDatabaseReader;
 use crate::track::TrackedSubstateValue;
