+++ radix-engine/src/vm/native_vm.rs
@@ -9,9 +9,9 @@ use crate::blueprints::transaction_processor::TransactionProcessorNativePackage;
 use crate::blueprints::transaction_tracker::TransactionTrackerNativePackage;
 use crate::errors::{NativeRuntimeError, RuntimeError, VmError};
 use crate::kernel::kernel_api::{KernelNodeApi, KernelSubstateApi};
-use crate::system::node_module::metadata::MetadataNativePackage;
-use crate::system::node_module::role_assignment::RoleAssignmentNativePackage;
-use crate::system::node_module::royalty::RoyaltyNativePackage;
+use crate::system::node_modules::metadata::MetadataNativePackage;
+use crate::system::node_modules::role_assignment::RoleAssignmentNativePackage;
+use crate::system::node_modules::royalty::RoyaltyNativePackage;
 use crate::system::system_callback::SystemLockData;
 use crate::types::*;
 use crate::vm::VmInvoke;
