+++ scrypto-unit/src/test_runner.rs
@@ -6,7 +6,7 @@ use std::process::Command;
 use radix_engine::blueprints::consensus_manager::*;
 use radix_engine::errors::*;
 use radix_engine::system::bootstrap::*;
-use radix_engine::system::node_module::type_info::TypeInfoSubstate;
+use radix_engine::system::node_modules::type_info::TypeInfoSubstate;
 use radix_engine::system::system::{FieldSubstate, KeyValueEntrySubstate};
 use radix_engine::system::system_db_reader::SystemDatabaseReader;
 use radix_engine::transaction::{
